flaskapp/requirements.txt
flask==2.0.1
requests==2.26.0
beautifulsoup4==4.10.0
Flask-Alembic==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-CLI==0.4.0
Flask-Migrate==3.1.0
# rpa==1.45.0
selenium==3.141.0
Pillow==8.3.2